Update README for imgix docs
Hi @sjelfull, I'm one of the front-end engineers over at imgix. We're currently working on a project to update the [Libraries page](https://docs.imgix.com/libraries) on our Docs site, to pull in more relevant information about our various libraries and 3rd-party plugins. Part of this project involves pulling README data directly from GitHub to display the detail page for each library, so we've been going through and [cleaning up the README files for our entire SDK](https://github.com/pulls?utf8=%E2%9C%93&q=is%3Aclosed+is%3Apr+user%3Aimgix+is%3Apublic+%22docs%3A%22+readme+author%3Asherwinski). The new detail page design looks like this: https://a.cl.ly/bLuGbnGX

We'd also like to pull in content directly from README files for the 3rd-party plugins that we link to, if the README file is in good enough shape. We think the content in this file would be great to include, but we do have to make a few little tweaks in order to make it fit perfectly. The changes proposed in this PR cover the small amount of clean-up we'd like to see before we start pulling this content into our docs site directly.

- I've replaced "Imgix" with "imgix" in a bunch of places, just to match the official spelling of the product
- I wrapped the top section in `<!-- ix-docs-ignore -->` tags. These get ignored by GitHub, but the content inside this wrapper will be stripped out when we include the content in our own docs. This essentially lets us avoid including the same information (project name, description, sometimes build badges) twice at the top of every page.
- I went ahead and added a table of contents, just to spruce it up a bit.
- And I fixed a couple of bullet-nested codeblocks that GitHub's renderer handles just fine but other Markdown renderers choke on.

The updated Docs page for this plug-in will of course include a link to this repo, plus a link directly to superbig.co. Let me know what you think!